fix for issue #1523
fix(paginationLinks): normalize `params` struct to query string early

Moved struct-to-string conversion of `arguments.params` to the top of the `paginationLinks()` function. This helps developer to pass a complete struct of variables instead of all variable separately.

Introduced new helper function `paramsToQueryString()` to convert a struct to a query string format using `encodeForUrl()`.