fix for issue #1217
fix(associations): fixed the issue for two foreignKeys to same table.

Refactor alias handling logic for join queries with dual foreign keys.

Now working fine for both first table and second table too.

Before it was not working because the first table was overriding the values of second table.

Now changed it according to the association keys used for joins defined in BelongsTo function.