Fix for running single tests
Updated the runner.cfm file for the app level tests to handle if the value passed to the format url attribute in html
Removed the "application.wo." prefix from the populate.cfm file as it is no longer necessary
Updated the html.cfm file to accept a parameter named type and set the values for the UI based on if the app level tests are run or the core level. Also updated the links for single tests so clicking on them runs the single test and clicking on the bundle name runs all the tests in that bundle.
Passed the value for the type variable from the runner.cfm files for the app level and core level tests.