Prevent for-index variable leakage.
The missing `var`|`local.` in `$splitOutsideFunctions()` was causing a race condition when trying to use a `model()` call inside an asynchronous iteration. The lack of local scoping was causing the concurrent threads to clobber each other's SQL statements. When I was looking for this issue, I found few others that looked suspicious.