Updating migratorversions
Updated the name of the new column from level_id to core_level
Updated the table name from migratorversions to _c_o_r_e_migrator_versions.
Added a fallback mechanism so that it checks if the table migratorversions exists and if it doesn then it renames that to _c_o_r_e_migrator_versions and also adds the new column core_level in it. If the table migratorversions doesn't already exist then it simply creates the new table _c_o_r_e_migrator_versions.
Updated the test cases to use the new migrator table name