Test cases fix for oracle
Updated the typeToSQL function so that all the options are passed in the options struct and checked if they exist first. This fixed the issue of oracle using the max length for column types instead of the provided one.
Updated the Base.cfc file and removed the addition of ";" for oracle to the queries as oracle does not accept it.
Added cfsqltype for datetime datatype.
Updated the populate.cfm file so that varchar2 is used for oracle instead of char