Fix for lucee5
Updated the tests.yml to run all test jobs and also removed unnecessary steps and added extra condition to not run cfpm for adobe 2018.
Added the $convertMaxRowsToLimit function to the $querySetup in Oracle adapter.
Added condition in $executeQuery function in Base.cfc to check if DATASOURCE exists in the queryAttributes argument then use that.