Update BLOB to TEXT and skip alter table tests
The coldfusion jdbc does not allow native array byte data insertion into SQLite so updated the datatype to TEXT and converted the data to base64 before inserting for SQLite.

Skipped the test for adding foreign key constraint for SQLite because it does not allow adding or removing constraints after the table has been created for SQLite.